+++ console/cypress/integration/data/modify/spec.ts
@@ -4,6 +4,7 @@ import {
   getTableName,
   getColName,
   getElementFromAlias,
+  createUntrackedFunctionSQL,
 } from '../../../helpers/dataHelpers';
 
 import {
@@ -11,10 +12,29 @@ import {
   validateCT,
   validateColumn,
   ResultType,
+  createFunctionRequest,
 } from '../../validators/validators';
 import { setPromptValue } from '../../../helpers/common';
 
 const testName = 'mod';
+
+export const passMTCreateFunctions = () => {
+  const tableName = getTableName(0, testName);
+  createFunctionRequest(createUntrackedFunctionSQL(`${tableName}_id_fn`, tableName), ResultType.SUCCESS);
+  cy.wait(5000);
+};
+
+export const passMTFunctionList = () => {
+  cy.get(getElementFromAlias('modify-table-edit-computed-field-0')).click();
+
+  cy.get(getElementFromAlias('functions-dropdown')).click();
+
+  cy.get('[data-test^="data_test_column_type_value_"]').should('have.length', 1);
+
+  cy.get('[data-test^="data_test_column_type_value_"]')
+      .first()
+      .should('have.text', `${getTableName(0, testName)}_id_fn`.toLowerCase());
+};
 
 export const passMTCreateTable = () => {
   cy.get(getElementFromAlias('data-create-table')).click();
@@ -291,6 +311,8 @@ export const passMTDeleteTable = () => {
   setPromptValue(getTableName(0, testName));
   cy.get(getElementFromAlias('delete-table')).click();
   cy.window().its('prompt').should('be.called');
+  cy.wait(5000);
+  cy.get('.notification-action-button').click();
   cy.wait(5000);
   cy.url().should('eq', `${baseUrl}/data/schema/public`);
   validateCT(getTableName(0, testName), ResultType.FAILURE);