+++ console/src/components/Services/Data/TableModify/ComputedFieldsEditor.js
@@ -311,11 +311,16 @@ const ComputedFieldsEditor = ({
                 Create new
               </RawSqlButton>
             </div>
-            <div className={styles.wd50percent}>
+            <div
+              className={styles.wd50percent}
+              data-test={'functions-dropdown'}
+            >
               <SearchableSelectBox
                 options={getSchemaFunctions(
                   functions,
-                  computedFieldFunctionSchema
+                  computedFieldFunctionSchema,
+                  table.table_name,
+                  table.table_schema
                 ).map(fn => getFunctionName(fn))}
                 onChange={handleFnNameChange}
                 value={computedFieldFunctionName}